bot/requirements.txt
discord.py
python-dotenv
requests
prometheus-client